updated the data type in ClientMessage to double to be complatible with Qt's Json library

+++ host/libs/StrataComm/src/ClientMessage.h
@@ -15,7 +15,7 @@ struct ClientMessage {
 
     QString handlerName;
     QJsonObject payload;
-    size_t messageID;
+    double messageID;
     QByteArray clientID;
     MessageType messageType;
 };
